Add files via upload
my name is ahmed bahgat , this my work
 module: print_direct
his  module is a feature that can be added to any module — in this case, we chose the Accounting module. It adds a button to the screen that, when clicked, sends the report directly to the printer and downloads it to the device, without displaying the report on screen.